Splitting the tests up into different files. This is a preliminary commit so I can do other things, will come back and fix things soon.
